requirements.txt
python-telegram-bot==22.1
redis==5.0.8
gunicorn==22.0.0